feat(moq-mux): CatalogExt trait + Catalog<E> for ergonomic catalog extensions
Instead of every extension hand-rolling a flatten wrapper plus AsRef/Deref impls,
provide that once as a library type. An application now writes only its sections
and an empty marker impl:

    #[derive(Serialize, Deserialize, Clone, Default)]
    struct Scte35Ext { scte35: Option<Scte35> }
    impl CatalogExt for Scte35Ext {}

    let mut p = Producer::with_catalog(&mut bc, Catalog::<Scte35Ext>::default())?;
    p.lock().video.renditions.insert(name, cfg); // base media, flat (deref)
    p.lock().ext.scte35 = Some(scte35);          // extension section

`Catalog<E>` flattens the base hang::Catalog and the extension E into one JSON
object on the wire (serde unions both), derefs to the base so the media fields
stay reachable directly (no `.base.` nesting), and impls AsRef for the MSF
side-channel. It plugs into the existing generic Producer/Consumer unchanged;
`NoExt` is the base-only default. Replaces the producer.rs extension test.
